refactor: merge adapter-cloudflare-writes tests into adapter-cloudflare
The two test suites tested complementary aspects of the same CF adapter
with nearly identical infrastructure. Combining them into a single suite
reduces duplication and simplifies maintenance.
