--- tests/integration/src/tests/adapter-cloudflare-writes/adapter-cloudflare-writes.test.ts
@@ -1,199 +1,0 @@
-import path from 'node:path'
-import { fileURLToPath } from 'node:url'
-import {
-  Duration,
-  Effect,
-  FetchHttpClient,
-  HttpClient, HttpClientRequest,
-  HttpClientResponse,
-  Layer,
-  Schema
-} from '@livestore/utils/effect'
-import { nanoid } from '@livestore/utils/nanoid'
-import { PlatformNode } from '@livestore/utils/node'
-import { Vitest } from '@livestore/utils-dev/node-vitest'
-import { expect } from 'vitest'
-
-const testDir = path.dirname(fileURLToPath(import.meta.url))
-const fixturesDir = path.join(testDir, 'fixtures')
-const testTimeout = Duration.toMillis(Duration.seconds(45))
-
-delete process.env.HTTP_PROXY
-delete process.env.http_proxy
-delete process.env.HTTPS_PROXY
-delete process.env.https_proxy
-delete process.env.ALL_PROXY
-delete process.env.all_proxy
-
-const { WranglerDevServerService } = await import('@livestore/utils-dev/wrangler')
-
-const withTestCtx = Vitest.makeWithTestCtx({
-  timeout: testTimeout,
-  makeLayer: () =>
-    Layer.mergeAll(
-      WranglerDevServerService.Default({
-        cwd: fixturesDir,
-        readiness: { connectTimeout: Duration.seconds(15) },
-      }).pipe(
-        Layer.provide(Layer.mergeAll(PlatformNode.NodeContext.layer, FetchHttpClient.layer)),
-      ),
-      FetchHttpClient.layer,
-    ),
-})
-
-const makeStoreHelpers = (serverUrl: string, storeId: string) =>
-  Effect.gen(function* () {
-    const client = (yield* HttpClient.HttpClient).pipe(
-      HttpClient.mapRequest((req) =>
-        req.pipe(
-          HttpClientRequest.prependUrl(serverUrl),
-          HttpClientRequest.setUrlParam('storeId', storeId),
-        ),
-      ),
-      HttpClient.filterStatusOk,
-    )
-
-    return {
-      createTodo: (id: string, title: string) =>
-        HttpClientRequest.post('/store/todos').pipe(
-          HttpClientRequest.bodyJson({ id, title }),
-          Effect.flatMap(client.execute),
-          Effect.flatMap(HttpClientResponse.schemaBodyJson(Schema.Struct({ id: Schema.String })))
-        ),
-
-      listTodos: () => client.get('/store/todos').pipe(
-        Effect.flatMap(HttpClientResponse.schemaBodyJson(Schema.Array(Schema.Struct({ id: Schema.String, title: Schema.String })))),
-      ),
-
-      getMetrics: () => client.get('/store/metrics').pipe(
-        Effect.flatMap(HttpClientResponse.schemaBodyJson(Schema.Struct({ totalRowsWritten: Schema.Number, totalRowsRead: Schema.Number }))),
-      ),
-
-      resetMetrics: () => client.del('/store/metrics').pipe(
-        Effect.flatMap(HttpClientResponse.schemaBodyJson(Schema.Struct({ totalRowsWritten: Schema.Number, totalRowsRead: Schema.Number }))),
-      ),
-
-      shutdownStore: () => client.post('/store/shutdown').pipe(
-        Effect.flatMap(HttpClientResponse.schemaBodyJson(Schema.Struct({ ok: Schema.Boolean }))),
-      ),
-    }
-  })
-
-Vitest.describe('adapter-cloudflare-writes', { timeout: testTimeout }, () => {
-  Vitest.live('rows written is below 20 per todo created', (test) =>
-    Effect.gen(function* () {
-      const server = yield* WranglerDevServerService
-      const storeId = `cf-writes-steady-state-${nanoid(6)}`
-      const { createTodo, listTodos, getMetrics, resetMetrics } = yield* makeStoreHelpers(server.url, storeId)
-
-      // Boot the store and discard initial write overhead so we measure steady-state only.
-      yield* createTodo('boot-todo', 'initial boot')
-      yield* resetMetrics()
-
-      const todos = Array.from({ length: 10 }, (_, i) => ({ id: `todo-${i}`, title: `item ${i}` }))
-      yield* Effect.forEach(todos, ({ id, title }) => createTodo(id, title), { concurrency: 1 })
-
-      const steadyStateMetrics = yield* getMetrics()
-      const writesPerTodo = steadyStateMetrics.totalRowsWritten / todos.length
-
-      const allTodos = yield* listTodos()
-      expect(allTodos).toHaveLength(todos.length + 1)
-      expect(writesPerTodo).toBeGreaterThan(0)
-      expect(writesPerTodo).toBeLessThan(20)
-
-      yield* Effect.promise(() => test.annotate(`${writesPerTodo.toFixed(1)} writes/todo (total: ${steadyStateMetrics.totalRowsWritten} for ${todos.length} todos)`))
-    }).pipe(withTestCtx(test)),
-  )
-
-  Vitest.live('cold start reopens persisted VFS state with zero writes', (test) =>
-    Effect.gen(function* () {
-      const server = yield* WranglerDevServerService
-      const storeId = `cf-writes-snapshot-restore-${nanoid(6)}`
-      const { createTodo, listTodos, getMetrics, resetMetrics, shutdownStore } = yield* makeStoreHelpers(
-        server.url,
-        storeId,
-      )
-
-      const todos = Array.from({ length: 5 }, (_, i) => ({ id: `todo-${i}`, title: `item ${i}` }))
-      yield* Effect.forEach(todos, ({ id, title }) => createTodo(id, title), { concurrency: 1 })
-
-      const preShutdownTodos = yield* listTodos()
-      expect(preShutdownTodos).toHaveLength(todos.length)
-
-      yield* shutdownStore()
-      yield* resetMetrics()
-
-      const postRestartTodos = yield* listTodos()
-      expect(postRestartTodos).toHaveLength(todos.length)
-
-      expect(postRestartTodos.map((t) => t.id)).toEqual(expect.arrayContaining(todos.map((t) => t.id)))
-
-      const restartMetrics = yield* getMetrics()
-
-      // Cold start with VFS-backed state should be cheap — just reopens the VFS.
-      // The eventlog is on native DO SQLite (1 row per event).
-      expect(restartMetrics.totalRowsRead).toBeGreaterThan(0)
-      expect(restartMetrics.totalRowsRead).toBeLessThan(50)
-      expect(restartMetrics.totalRowsWritten).toBe(0)
-
-      yield* Effect.promise(() => test.annotate(`${restartMetrics.totalRowsWritten} writes, ${restartMetrics.totalRowsRead} reads on cold start`))
-    }).pipe(withTestCtx(test)),
-  )
-
-  Vitest.live('data survives multiple shutdown cycles', (test) =>
-    Effect.gen(function* () {
-      const server = yield* WranglerDevServerService
-      const storeId = `cf-writes-multi-cycle-${nanoid(6)}`
-      const { createTodo, listTodos, shutdownStore } = yield* makeStoreHelpers(server.url, storeId)
-
-      yield* createTodo('todo-a', 'first cycle')
-      yield* shutdownStore()
-
-      yield* createTodo('todo-b', 'second cycle')
-      yield* shutdownStore()
-
-      yield* createTodo('todo-c', 'third cycle')
-
-      const allTodos = yield* listTodos()
-      expect(allTodos).toHaveLength(3)
-      expect(allTodos).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({ id: 'todo-a', title: 'first cycle' }),
-          expect.objectContaining({ id: 'todo-b', title: 'second cycle' }),
-          expect.objectContaining({ id: 'todo-c', title: 'third cycle' }),
-        ]),
-      )
-    }).pipe(withTestCtx(test)),
-  )
-
-  Vitest.live('rows written is below 20 per todo created after cold start', (test) =>
-    Effect.gen(function* () {
-      const server = yield* WranglerDevServerService
-      const storeId = `cf-writes-post-restart-${nanoid(6)}`
-      const { createTodo, listTodos, getMetrics, resetMetrics, shutdownStore } = yield* makeStoreHelpers(
-        server.url,
-        storeId,
-      )
-
-      yield* createTodo('seed-todo', 'seed')
-      yield* shutdownStore()
-
-      // Reboot — VFS-backed state persists automatically
-      yield* createTodo('post-restart-boot', 'boot after restart')
-      yield* resetMetrics()
-
-      const todos = Array.from({ length: 5 }, (_, i) => ({ id: `post-restart-${i}`, title: `item ${i}` }))
-      yield* Effect.forEach(todos, ({ id, title }) => createTodo(id, title), { concurrency: 1 })
-
-      const metrics = yield* getMetrics()
-      const writesPerTodo = metrics.totalRowsWritten / todos.length
-
-      const allTodos = yield* listTodos()
-      expect(allTodos).toHaveLength(todos.length + 2) // seed + boot + N
-      expect(writesPerTodo).toBeGreaterThan(0)
-      expect(writesPerTodo).toBeLessThan(20)
-
-      yield* Effect.promise(() => test.annotate(`${writesPerTodo.toFixed(1)} writes/todo after cold start`))
-    }).pipe(withTestCtx(test)),
-  )
-})
