--- tests/integration/src/tests/adapter-cloudflare-writes/fixtures/wrangler.toml
@@ -1,16 +1,0 @@
-name = "adapter-cloudflare-writes-test"
-main = "./worker.ts"
-compatibility_date = "2025-08-24"
-compatibility_flags = ["enable_request_signal"] # Needed for HTTP RPC streams
-
-[[durable_objects.bindings]]
-name = "SYNC_BACKEND_DO"
-class_name = "SyncBackendDO"
-
-[[durable_objects.bindings]]
-name = "TEST_STORE_DO"
-class_name = "WriteTrackingStoreDo"
-
-[[migrations]]
-tag = "v1"
-new_sqlite_classes = ["SyncBackendDO", "WriteTrackingStoreDo"]
