--- tests/integration/src/tests/adapter-cloudflare-writes/schema.ts
@@ -1,25 +1,0 @@
-import { Events, makeSchema, State } from '@livestore/common/schema'
-import { Schema } from '@livestore/utils/effect'
-
-const todos = State.SQLite.table({
-  name: 'todos',
-  columns: {
-    id: State.SQLite.text({ primaryKey: true }),
-    title: State.SQLite.text(),
-  },
-})
-
-export const events = {
-  todoCreated: Events.synced({
-    name: 'todoCreated',
-    schema: Schema.Struct({ id: Schema.String, title: Schema.String }),
-  }),
-}
-
-const materializers = State.SQLite.materializers(events, {
-  todoCreated: ({ id, title }) => todos.insert({ id, title }),
-})
-
-export const tables = { todos }
-const state = State.SQLite.makeState({ tables, materializers })
-export const schema = makeSchema({ state, events })
