--- tests/integration/src/tests/adapter-cloudflare-writes/vitest.config.ts
@@ -1,5 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: { server: { deps: { inline: ['@effect/vitest'] } } },
-})
